fix(scan): remove cycles from `ballot-interpreter-nh`
